+++ src/shaders/image_effects.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+extern const char *k_image_frag_shader;